【修改描述】缓存盘空间划分 【DTS/CR/AR单号】AR20240119615093 【合入版本】BeiMing-K 24.0.0,
Match-id-ebfeb4779ae36dc0cdf35362af2db70cc592110b

+++ src/sdk/bio.h
@@ -64,7 +64,7 @@ public:
      * @param[out]: location: location info
      * @return: return RETURN_CACHE_OK mean success, others, return non-zero value
      */
-    CResult CalculateLocation(uint64_t objectId, ObjLocation &location);
+    CResult CalculateLocation(uint32_t objectId, ObjLocation &location);
 
     /* *
      * @brief: Put value
